floats and mixers

+++ Assets/Resources/Scripts/Thermometer.cs
@@ -18,7 +18,7 @@ public class Thermometer : Device {
 
 	public override void set(Day d) {
 		
-		int temperature = d.Temperature;
+		float temperature = d.Temperature;
 		if (temperature < 20) {
 			orange.moveToTop ();
 			green.moveToTop ();
